+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.14.0-rc.59",
+  sdkVersion: "3.14.0-rc.60",
   genVersion: "2.730.5",
-  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.59 2.730.5 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.14.0-rc.60 2.730.5 2.0 @orq-ai/node",
 } as const;