+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -204,7 +204,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-27T08:59:46.066Z",
+    "2025-10-27T09:34:54.012Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -244,7 +244,7 @@ export const RetrieveDatasetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-27T08:59:46.066Z"))
+  updated: z.date().default(() => new Date("2025-10-27T09:34:54.012Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {