+++ packages/orq-rc/RELEASES.md
@@ -4589,3 +4589,13 @@ Based on:
 - [typescript v3.14.0-rc.59] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.59] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.59 - packages/orq-rc
+
+## 2025-10-27 09:36:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.639.3 (2.730.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.60] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.60] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.60 - packages/orq-rc